Core SDK 2.1.2
- Added support for Odin core sdk version 2.1.2, which brings the Audio
  Events "Is Silent Changed" and "Positions Changed" for Encoders and
  Decoders
- Improved logging to avoid showing errors during expected code
  scenarios

+++ Source/Odin/Public/OdinNative/OdinNativeBlueprint.h
@@ -173,7 +173,7 @@ enum class EOdinAudioEvents : uint8 {
     /**
      * Enables all available audio events.
      */
-    AUDIO_EVENTS_ALL = -1 UMETA(DisplayName = "ALL"),
+    AUDIO_EVENTS_ALL = static_cast<uint8>(-1) UMETA(DisplayName = "ALL"),
 };
 inline bool operator==(int a, EOdinAudioEvents b)
 {
